LPS-61952 Move com.liferay.portal.model.adapter to com.liferay.portal.kernel.model.adapter
