+++ portal-service/src/com/liferay/portal/kernel/model/adapter/builder/packageinfo
